+++ attributes/dbconfig.rb
@@ -1,35 +1,55 @@
 # All the configuration files that can be dumped
 # the attribute-based-configuration
-# dump anything into default['mongodb']['config'][<setting>] = <value>
+# dump anything into default['mongodb']['config']['mongod'][<setting>] = <value>
 # these options are in the order of mongodb docs
 
 include_attribute 'sc-mongodb::default'
 
-default['mongodb']['config']['port'] = 27_017
-default['mongodb']['config']['bind_ip'] = '0.0.0.0'
-# Workaround for opscode/chef#1507, which prevents users from
-# unsetting our default with a nil override.
-# So we make sure to unset logpath when syslog is set since the two
-# settings are incompatible.
-# For more information see: edelight/chef-mongodb#310
-unless node['mongodb']['config']['syslog']
-  default['mongodb']['config']['logpath'] = '/var/log/mongodb/mongodb.log'
-end
-default['mongodb']['config']['logappend'] = true
-# The platform_family? syntax in attributes files was added in Chef 11
-# if node.platform_family?("rhel", "fedora") then
+# mongod defaults
+default['mongodb']['config']['mongod']['net']['port'] = 27017
+default['mongodb']['config']['mongod']['net']['bindIp'] = '0.0.0.0'
+
+default['mongodb']['config']['mongod']['systemLog']['destination'] = 'file'
+default['mongodb']['config']['mongod']['systemLog']['logAppend'] = true
+default['mongodb']['config']['mongod']['systemLog']['path'] = '/var/log/mongodb/mongod.log'
+
 case node['platform_family']
 when 'rhel', 'fedora'
-  default['mongodb']['config']['fork'] = true
-  default['mongodb']['config']['pidfilepath'] = '/var/run/mongodb/mongodb.pid'
-else
-  default['mongodb']['config']['fork'] = false
+  default['mongodb']['config']['mongod']['processManagement']['fork'] = true
+  default['mongodb']['config']['mongod']['processManagement']['pidFilePath'] = '/var/run/mongodb/mongod.pid'
 end
-default['mongodb']['config']['dbpath'] = '/var/lib/mongodb'
-default['mongodb']['config']['nojournal'] = false
-default['mongodb']['config']['rest'] = false
-default['mongodb']['config']['smallfiles'] = false
-default['mongodb']['config']['oplogSize'] = nil
 
-default['mongodb']['config']['replSet'] = nil
-default['mongodb']['config']['keyFile'] = '/etc/mongodb.key' if node['mongodb']['key_file_content']
+default['mongodb']['config']['mongod']['storage']['journal']['enabled'] = true
+default['mongodb']['config']['mongod']['storage']['dbPath'] = case node['platform_family']
+                                                              when 'rhel', 'fedora'
+                                                                '/var/lib/mongo'
+                                                              else
+                                                                '/var/lib/mongodb'
+                                                              end
+
+default['mongodb']['config']['mongod']['storage']['engine'] = 'wiredTiger'
+
+default['mongodb']['config']['mongod']['replication']['oplogSizeMB'] = nil
+default['mongodb']['config']['mongod']['replication']['replSetName'] = nil
+default['mongodb']['config']['mongod']['replication']['secondaryIndexPrefetch'] = nil
+default['mongodb']['config']['mongod']['replication']['enableMajorityReadConcern'] = nil
+
+default['mongodb']['config']['mongod']['security']['keyFile'] = nil
+
+# mongos defaults
+default['mongodb']['config']['mongos']['net']['port'] = 27017
+default['mongodb']['config']['mongos']['net']['bindIp'] = '0.0.0.0'
+
+default['mongodb']['config']['mongos']['systemLog']['destination'] = 'file'
+default['mongodb']['config']['mongos']['systemLog']['logAppend'] = true
+default['mongodb']['config']['mongos']['systemLog']['path'] = '/var/log/mongodb/mongos.log'
+
+case node['platform_family']
+when 'rhel', 'fedora'
+  default['mongodb']['config']['mongos']['processManagement']['fork'] = true
+  default['mongodb']['config']['mongos']['processManagement']['pidFilePath'] = '/var/run/mongodb/mongos.pid'
+end
+
+default['mongodb']['config']['mongos']['sharding']['configDB'] = nil
+
+default['mongodb']['config']['mongos']['security']['keyFile'] = nil