+++ metadata.rb
@@ -7,7 +7,6 @@ long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version           '0.18.1'
 
 recipe 'sc-mongodb', 'Installs and configures a single node mongodb instance'
-recipe 'sc-mongodb::10gen_repo', 'Adds the 10gen repo to get the latest packages'
 recipe 'sc-mongodb::mongos', 'Installs and configures a mongos which can be used in a sharded setup'
 recipe 'sc-mongodb::configserver', 'Installs and configures a configserver for mongodb sharding'
 recipe 'sc-mongodb::shard', 'Installs and configures a single shard'
@@ -17,13 +16,19 @@ recipe 'sc-mongodb::mms_backup_agent', 'Installs and configures a MongoDB MMS Ba
 
 depends 'apt', '>= 1.8.2'
 depends 'yum', '>= 3.0'
-depends 'python'
 depends 'build-essential', '>= 5.0.0'
 
-%w(ubuntu debian centos redhat amazon).each do |os|
+%w(
+  amazon
+  centos
+  debian
+  oracle
+  redhat
+  ubuntu
+).each do |os|
   supports os
 end
 
 source_url 'https://github.com/sous-chefs/mongodb' if respond_to?(:source_url)
 issues_url 'https://github.com/sous-chefs/mongodb/issues' if respond_to?(:issues_url)
-chef_version '>= 11.0' if respond_to?(:chef_version)
+chef_version '>= 12.1' if respond_to?(:chef_version)