+++ recipes/install.rb
@@ -1,10 +1,37 @@
-# install the mongodb_org_repo if necessary
-include_recipe 'sc-mongodb::mongodb_org_repo' if %w(10gen mongodb-org).include?(node['mongodb']['install_method'])
+#
+# Cookbook Name:: mongodb
+# Recipe:: install
+#
+# Copyright 2011, edelight GmbH
+# Authors:
+#       Markus Korn <markus.korn@edelight.de>
+#
+# Copyright 2016-2017, Grant Ridder
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 
+# install the mongodb org repo if necessary
+include_recipe 'sc-mongodb::mongodb_org_repo' if node['mongodb']['install_method'] == 'mongodb-org'
+
+# TODO: still need all of the tools?
+# yum_package[autoconf, bison, flex, gcc, gcc-c++, gettext, kernel-devel, make, m4, ncurses-devel, patch]
 build_essential 'build-tools'
 
+config_type = node['mongodb']['is_mongos'] ? 'mongos' : 'mongod'
+
 # prevent-install defaults, but don't overwrite
-file node['mongodb']['sysconfig_file'] do
+file node['mongodb']['sysconfig_file'][config_type] do
   content 'ENABLE_MONGODB=no'
   group node['mongodb']['root_group']
   owner 'root'
@@ -13,14 +40,15 @@ file node['mongodb']['sysconfig_file'] do
 end
 
 # just-in-case config file drop
-template node['mongodb']['dbconfig_file'] do
+template "#{node['mongodb']['dbconfig_file'][config_type]} install" do
+  path node['mongodb']['dbconfig_file'][config_type]
   cookbook node['mongodb']['template_cookbook']
-  source node['mongodb']['dbconfig_file_template']
+  source node['mongodb']['dbconfig_file']['template']
   group node['mongodb']['root_group']
   owner 'root'
   mode 0644
   variables(
-    config: node['mongodb']['config']
+    config: node['mongodb']['config'][config_type]
   )
   helpers MongoDBConfigHelpers
   action :create_if_missing
@@ -41,7 +69,8 @@ execute 'mongodb-systemctl-daemon-reload' do
   action :nothing
 end
 
-template init_file do
+template "#{init_file} install" do
+  path init_file
   cookbook node['mongodb']['template_cookbook']
   source node['mongodb']['init_script_template']
   group node['mongodb']['root_group']
@@ -49,31 +78,45 @@ template init_file do
   mode mode
   variables(
     provides: 'mongod',
-    dbconfig_file: node['mongodb']['dbconfig_file'],
-    sysconfig_file: node['mongodb']['sysconfig_file'],
+    dbconfig_file: node['mongodb']['dbconfig_file'][config_type],
+    sysconfig_file: node['mongodb']['sysconfig_file'][config_type],
     ulimit: node['mongodb']['ulimit'],
-    bind_ip: node['mongodb']['config']['bind_ip'],
-    port: node['mongodb']['config']['port'],
+    bind_ip: node['mongodb']['config'][config_type]['net']['bindIp'],
+    port: node['mongodb']['config'][config_type]['net']['port'],
     user: node['mongodb']['user']
   )
   action :create_if_missing
 
-  if platform_family?('rhel') && node['platform'] != 'amazon' && node['platform_version'].to_i >= 7
+  if (platform_family?('rhel') && node['platform'] != 'amazon' && node['platform_version'].to_i >= 7) || (node['platform'] == 'debian' && node['platform_version'].to_i >= 8)
     notifies :run, 'execute[mongodb-systemctl-daemon-reload]', :immediately
   end
 end
 
-# install
+# Adjust the version number for RHEL style if needed
+package_version = case node['platform_family']
+                  when 'rhel'
+                    if node['platform'] == 'amazon'
+                      "#{node['mongodb']['package_version']}-1.amzn1"
+                    else
+                      "#{node['mongodb']['package_version']}-1.el#{node['platform_version'].to_i}"
+                    end
+                  when 'fedora'
+                    "#{node['mongodb']['package_version']}-1.el7"
+                  else
+                    node['mongodb']['package_version']
+                  end
+
+# Install
 package node['mongodb']['package_name'] do
   options node['mongodb']['packager_options']
   action :install
-  version node['mongodb']['package_version']
+  version package_version
   not_if { node['mongodb']['install_method'] == 'none' }
 end
 
 # Create keyFile if specified
 if node['mongodb']['key_file_content']
-  file node['mongodb']['config']['keyFile'] do
+  file node['mongodb']['config'][config_type]['keyFile'] do
     owner node['mongodb']['user']
     group node['mongodb']['group']
     mode  '0600'
@@ -81,3 +124,5 @@ if node['mongodb']['key_file_content']
     content node['mongodb']['key_file_content']
   end
 end
+
+node.default['mongodb']['config'][config_type]['security']['keyFile'] = nil if node['mongodb']['key_file_content'].nil?