+++ test/integration/replicaset/inspec/replicaset_spec.rb
@@ -1,5 +1,5 @@
 # mongodb | mongod
-describe service('mongodb') do
+describe service('mongod') do
   it { should be_installed }
   it { should be_enabled }
   it { should be_running }